fix: add screenshot utils, fixing review
