+++ app/components/keyboard/modal/BrandForm.vue
@@ -77,6 +77,7 @@
 import slugify from 'slugify'
 import { z } from 'zod'
 import country from 'flag-icons/country.json'
+import { discordInviteRegex, instagramProfileRegex } from '~/utils'
 
 const emit = defineEmits(['onSuccess'])
 
@@ -99,9 +100,6 @@ const brand = ref({
   discord: '',
 })
 
-const discordInviteRegex = /discord\.gg\/[a-zA-Z0-9]+/
-const instagramRegex = /^(https?:\/\/)?(www\.)?instagram\.com\/[a-zA-Z0-9._-]+/
-
 const schema = z.object({
   name: z.string().min(1),
   slug: z
@@ -116,7 +114,7 @@ const schema = z.object({
   website: z.url().nullish().or(z.string().min(0).max(0)),
   instagram: z
     .url()
-    .regex(instagramRegex, 'Invalid Instagram profile URL')
+    .regex(instagramProfileRegex, 'Invalid Instagram profile URL')
     .nullish()
     .or(z.string().min(0).max(0)),
   discord: z
